LEXEVS-1903.  LEXEVS-1931.  Support for value sets with some modification of search extensions scope to get things working smoothly.
